+++ src/styles/globals.css
@@ -2,24 +2,6 @@
 
 @config '../../tailwind.config.ts';
 
-/*
-  The default border color has changed to `currentColor` in Tailwind CSS v4,
-  so we've added these compatibility styles to make sure everything still
-  looks the same as it did with Tailwind CSS v3.
-
-  If we ever want to remove these styles, we need to add an explicit border
-  color utility to any element that depends on these defaults.
-*/
-@layer base {
-  *,
-  ::after,
-  ::before,
-  ::backdrop,
-  ::file-selector-button {
-    border-color: var(--color-gray-200, currentColor);
-  }
-}
-
 @utility card {
   @apply bg-secondary rounded-xl p-6 shadow-xs dark:shadow-none;
   @apply print:break-inside-avoid print:bg-transparent print:p-0 print:shadow-none;
